feat: 졸업 요건 체크 및 파싱 기능 추가
