+++ app/services/checkers/base.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+from app.services.checkers.context import ValidationContext
+
+
+class RuleChecker(ABC):
+    @abstractmethod
+    def check(self, ctx: ValidationContext) -> None:
+        ...
